move jest config out to a preset and remove from package.json
