+++ package.json
@@ -53,10 +53,11 @@
     "glob": "^7.1.6",
     "husky": "^4.3.0",
     "jest": "^26.4.2",
-    "lerna": "3.22.1",
+    "lerna": "^3.22.1",
     "ts-jest": "^26.4.1",
     "tslint": "^6.1.3",
     "tslint-react": "^5.0.0",
     "typescript": "^3.9.7"
-  }
+  },
+  "name": "flex-plugin-builder"
 }
