--- packages/create-flex-plugin/templates/core/src/setupTests.js
@@ -1,8 +1,0 @@
-require('babel-polyfill');
-
-import { configure } from 'enzyme/build';
-import Adapter from 'enzyme-adapter-react-16/build';
-
-configure({
-  adapter: new Adapter(),
-});
