+++ packages/create-flex-plugin/templates/js/package.json
@@ -6,17 +6,12 @@
     "postinstall": "flex-plugin pre-script-check"
   },
   "dependencies": {
-    "flex-plugin": "{{flexPluginVersion}}",
+    "flex-plugin-scripts": "{{pluginScriptsVersion}}",
     "react": "16.5.2",
-    "react-dom": "16.5.2",
-    "react-emotion": "9.2.12"
+    "react-dom": "16.5.2"
   },
   "devDependencies": {
-    "@twilio/flex-ui": "{{flexSdkVersion}}",
-    "babel-polyfill": "^6.26.0",
-    "enzyme": "^3.10.0",
-    "enzyme-adapter-react-16": "^1.14.0",
-    "flex-plugin-scripts": "{{pluginScriptsVersion}}"
+    "@twilio/flex-ui": "{{flexSdkVersion}}"
   },
   "browserslist": {
     "production": [
