+++ packages/flex-plugin-scripts/package.json
@@ -38,6 +38,9 @@
     "test:watch": "cd ../.. && jest packages/flex-plugin-scripts --watch"
   },
   "dependencies": {
+    "flex-plugin": "4.0.2-beta.0",
+    "react-emotion": "9.2.12",
+    "flex-plugin-test": "4.0.8-beta.0",
     "@babel/core": "^7.11.6",
     "@babel/preset-env": "^7.11.5",
     "@babel/preset-react": "^7.10.4",
@@ -48,23 +51,18 @@
     "@k88/typescript-compile-error-formatter": "^1.0.3",
     "babel-loader": "^8.1.0",
     "babel-plugin-named-asset-import": "^0.3.6",
-    "babel-preset-react-app": "^9.1.2",
     "css-loader": "^3.6.0",
     "esm": "^3.2.25",
     "flex-dev-utils": "^4.0.7-beta.0",
     "fork-ts-checker-webpack-plugin": "^4.1.6",
     "form-data": "^2.5.1",
     "html-webpack-plugin": "^4.5.0",
-    "jest-environment-jsdom-fourteen": "^1.0.1",
-    "jest-watch-typeahead": "^0.6.1",
     "node-ipc": "^9.1.1",
     "pnp-webpack-plugin": "^1.6.4",
     "postcss-flexbugs-fixes": "^4.2.1",
     "postcss-loader": "^3.0.0",
     "postcss-preset-env": "^6.7.0",
     "querystring": "^0.2.0",
-    "react-app-polyfill": "^1.0.6",
-    "react-test-renderer": "^16.13.1",
     "resolve-url-loader": "^3.1.1",
     "sass-loader": "^8.0.2",
     "style-loader": "^1.2.1",
@@ -75,14 +73,10 @@
     "webpack-dev-server": "^3.11.0"
   },
   "devDependencies": {
-    "@jest/types": "^26.3.0",
     "@types/node-ipc": "^9.1.3",
     "@types/terser-webpack-plugin": "^2.2.1",
     "@types/webpack": "^4.41.22",
-    "@types/webpack-dev-server": "^3.11.0",
-    "babel-jest": "^26.3.0",
-    "jest": "^26.4.2",
-    "react-test-renderer": "^16.13.1"
+    "@types/webpack-dev-server": "^3.11.0"
   },
   "peerDependencies": {
     "flex-plugin": "*"
